requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
torchvision==0.15.2
opencv-python-headless==4.8.1.78
pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
scikit-image==0.21.0
python-dotenv==1.0.0
gunicorn==21.2.0